move ggml_sycl_fattn_* into fattn-buffers.hpp
